Added tests and error throwing to services

+++ app/scripts/services/itemFactory.js
@@ -38,34 +38,38 @@ angular.module('svr2App')
 	@return {object} items
 	**/
 	factory.addItem = function(id){
-		var item = {
-			id : id,
-            ranges: rangeService.getRanges(),
-            addRange: function(start, stop){
-                rangeService.addRange(item.ranges, start, stop);
-                return item;
-            },
-            removeRange: function(start){
-                rangeService.removeRange(item.ranges, start);
-                return item;
-            },
-            updateRange: function updateRange(oldStart, start, stop){
-                rangeService.updateRange(item.ranges, oldStart, start, stop)
-                return item;
-            },
-            getRangeIndex: function(start){
-                return rangeService.getRangeIndex(item.ranges, start);
-            },
-            getRange: function getRange(start){
-                return rangeService.getRange(item.ranges, start);
-            },
-            hasRange: function hasRange(currentPosition){
-                return rangeService.hasRange(item.ranges, currentPosition);
-            }
-		}
+        if(!factory.getItem(id)){
+    		var item = {
+    			id : id,
+                ranges: rangeService.getRanges(),
+                addRange: function(start, stop){
+                    rangeService.addRange(item.ranges, start, stop);
+                    return item;	
+                },
+                removeRange: function(start){
+                    rangeService.removeRange(item.ranges, start);
+                    return item;
+                },
+                updateRange: function updateRange(oldStart, start, stop){
+                    rangeService.updateRange(item.ranges, oldStart, start, stop)
+                    return item;
+                },
+                getRangeIndex: function(start){
+                    return rangeService.getRangeIndex(item.ranges, start);
+                },
+                getRange: function getRange(start){
+                    return rangeService.getRange(item.ranges, start);
+                },
+                hasRange: function hasRange(currentPosition){
+                    return rangeService.hasRange(item.ranges, currentPosition);
+                }
+    		}
 
-		items[item.id] = item;
-
+    		items[item.id] = item;
+        }
+        else{
+            throw {name : "ItemError", message : "Item already exist. Can't add item."}; 
+        }    
         return factory;
 	}
